moved the get_channel() method to the conversations table
